More efficient JSON serialization of the error
